add jwt

+++ resolvers/mutation.js
@@ -3,19 +3,27 @@ import jwt from "jsonwebtoken";
 import bcrypt from "bcrypt";
 import { AuthenticationError, ForbiddenError } from "apollo-server-express";
 import * as dotenv from "dotenv";
+import mongoose from "mongoose";
 
 dotenv.config();
 
 const JWT_KEY = process.env.JWT_SECRET_KEY;
 
 export const Mutation = {
-  newNote: async (_, args) => {
+  newNote: async (_, { title, content }, { user }) => {
+    if (!user) {
+      throw new ForbiddenError("User not logged in");
+    }
     return await models.Note.create({
-      title: args.title,
-      content: args.content,
+      title: title,
+      content: content,
+      author: new mongoose.Types.ObjectId(user.id),
     });
   },
-  updateNote: async (_, { id, title, content }) => {
+  updateNote: async (_, { id, title, content }, { user }) => {
+    if (!user) {
+      throw new ForbiddenError("User not logged in");
+    }
     return await models.Note.findByIdAndUpdate(
       id,
       {
@@ -29,7 +37,10 @@ export const Mutation = {
       }
     );
   },
-  deleteNote: async (_, { id }) => {
+  deleteNote: async (_, { id }, { user }) => {
+    if (!user) {
+      throw new ForbiddenError("User not logged in");
+    }
     return await models.Note.findByIdAndDelete(id);
   },
   signUp: async (_, { username, email, password }) => {
